% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.MAESPA.R
\name{write.config.MAESPA}
\alias{write.config.MAESPA}
\title{Write MAESPA configuration files}
\usage{
write.config.MAESPA(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}
}
\value{
configuration file for MAESPA for given run
}
\description{
Writes a config file for Maespa
}
\details{
Requires a pft xml object, a list of trait values for a single model run,
and the name of the file to create
}
\author{
Tony Gardella
}
